# api/openapi.yaml
openapi: 3.1.0
info:
  title: Nuvex API
  version: 0.1.0
  description: >
    Read model for operators and dashboards. Responses are not protocol truth.
    Finalized results come from Solana accounts. GET /health returns authority: none.
    Requests, nodes, and network counts are served from the indexer store.
    Jobs, models, and prices stay unimplemented.
paths:
...
  /v1/requests:
    get:
      summary: Indexed requests
      responses:
        "200":
          description: Indexed request accounts. An empty list means none were observed.
        "503":
          description: No read model is configured.
  /v1/requests/{id}:
    get:
      summary: One indexed request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The indexed row.
        "404":
          description: No indexed row. The account may still exist on chain.
        "503":
          description: No read model is configured.
...